requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
requests
SQLAlchemy
arrow
Flask-Login
Flask-Mobility
python-dotenv
alembic
psycopg2-binary
oauthlib
flake8
qrcode